Build subscriber-app and simulator
