+++ .gitignore
@@ -2,3 +2,6 @@ external-scaler-containerapps-and-k8s/metricsReaders/azureMetricsReader_testgo
 .env
 
 __pycache__
+
+
+.generated-keys.json
