+++ external-scaler-containerapps-and-k8s/infra/.gitignore
@@ -1,0 +1,1 @@
+simulator
