+++ external-scaler-containerapps-and-k8s/infra/bicep/.simulator_tag
@@ -1,0 +1,1 @@
+v0.5
