+++ external-scaler-containerapps-and-k8s/infra/bicep/sample.env
@@ -1,0 +1,4 @@
+# Copy this to .env and fill in the values
+RG_NAME=YOUR_RESOURCE_GROUP_NAME
+LOCATION=YOUR_LOCATION
+BASE_NAME=test1
