+++ subscriber-app/config.py
@@ -10,7 +10,8 @@ SERVICE_BUS_TOPIC_NAME = os.getenv("SERVICE_BUS_TOPIC_NAME", "")
 SERVICE_BUS_SUBSCRIPTION_NAME = os.getenv("SERVICE_BUS_SUBSCRIPTION_NAME", "")
 
 if not SERVICE_BUS_NAMESPACE and not SERVICE_BUS_CONNECTION_STRING:
-    raise ValueError("One of SERVICE_BUS_NAMESPACE or SERVICE_BUS_CONNECTION_STRING must be set")
+    raise ValueError(
+        "One of SERVICE_BUS_NAMESPACE or SERVICE_BUS_CONNECTION_STRING must be set")
 if not SERVICE_BUS_TOPIC_NAME:
     raise ValueError("SERVICE_BUS_TOPIC_NAME must be set")
 if not SERVICE_BUS_SUBSCRIPTION_NAME:
